Habitverse/HabitVerse-V3-main/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
Jinja2==3.1.4
alembic==1.13.2
psycopg2-binary==2.9.9  # optional, only if using Postgres
PyMySQL==1.1.1  # MySQL driver for SQLAlchemy (mysql+pymysql)